// src/app/pages/models/kycdata.model.ts
export interface kycData {
            id: string;
            kycid: string;                
            message: string; 
            kycemail: string; 
            kycdone: string;
       name: string;
       loginphone: string;
       kycphone: string;
       googleid: string;
       loginemail: string;
       updatedate: string;
       country: string;
       address: string;
       kycdoctype: string;
       kycdocnumber: string;
       kycdocument: string;

};